requirements 2.txt
admin==0.0.1
aiodns==3.0.0
aiohttp==3.8.1
aioredis==2.0.0
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.2
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.1
asyncio==3.4.3
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
boto==2.49.0
Brotli==1.0.9
brython==3.10.3
bs4==0.0.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
commonmark==0.9.1
cryptography==35.0.0
cssselect==1.1.0
dateparser==1.1.0
decorator==5.1.0
deepdiff==5.6.0
Deprecated==1.2.13
device-detector==0.10
distlib==0.3.3
Django==3.2.9
django-extensions==3.1.5
dotty-dict==1.3.0
elasticsearch==7.13.4
execnet==1.9.0
fake-useragent==0.1.11
fasteners==0.16.3
filelock==3.4.0
frozenlist==1.2.0
geckodriver-autoinstaller==0.1.0
geoip2==4.2.0
geolite2utils==1.0.4
gssapi==1.7.2
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.29.0
jedi==0.18.1
jsoncodable==0.1.7
jsonpickle==2.0.0
k-selenium-cookies==0.0.4
keras==2.7.0
kproxy==0.0.1
krb5==0.2.0
lark==1.0.0
lxml==4.6.4
matplotlib-inline==0.1.3
maxminddb==2.2.0
maxminddb-geolite2==2018.703
more-itertools==8.11.0
multidict==5.2.0
noraise==0.0.16
nose==1.3.7
ordered-set==4.0.2
outcome==1.1.0
packaging==21.3
parameterized==0.8.1
parse==1.19.0
parso==0.8.2
path==16.2.0
pdfminer.six==20211012
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
prodict==0.8.18
prompt-toolkit==3.0.22
ptyprocess==0.7.0
py==1.11.0
pycares==4.1.2
pycparser==2.21
pydantic==1.8.2
pyee==8.2.2
Pygments==2.10.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyotp==2.6.0
pyparsing==3.0.6
pyppeteer==0.2.6
pyquery==1.4.3
pyspnego==0.3.1
pytest==6.2.5
pytest-forked==1.3.0
pytest-html==2.0.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.4.0
python-dateutil==2.8.2
python-geoip==1.2
python-geoip-geolite2==2015.303
python-geoip-geolite2-coex==2018.1113
python-geoip-geolite2-yplan==2019.1224
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.0.1
regex==2021.11.10
requests==2.26.0
requests-core==0.0.0
requests-file==1.5.1
requests-futures==1.0.0
requests-html==0.10.0
requests-kerberos==0.13.0
requests-xml==0.2.3
rich==10.14.0
sbvirtualdisplay==1.0.0
selenium==4.0.0
selenium-browser==0.0.15
selenium-chrome-screenshot==0.0.1
selenium-firefox==2.0.7
seleniumbase==2.1.9
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tensorflow-estimator==2.7.0
tldextract==3.1.2
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
tracardi==0.6.19
tracardi-dot-notation==0.6.5
tracardi-graph-runner==0.6.6
tracardi-maxmind-geolite2==0.6.0
tracardi-plugin-sdk==0.6.25
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.10.0
w3lib==1.22.0
Wand==0.6.7
wcwidth==0.2.5
websockets==9.1
wget==3.2
wrapt==1.13.3
wsproto==1.0.0
xmljson==0.2.1
xpath-utils==0.0.3
yarl==1.7.2
zipp==3.6.0
